requirements.txt
# Core API
fastmcp
fastapi==0.110.0
uvicorn[standard]==0.27.1

# HTTP client for calling MCP services
httpx==0.27.0

# DB (SQLite) + utils
sqlalchemy==2.0.29
sqlite-utils==3.36
sqlmodel==0.0.14

# Scheduling (if mesh-core runs cron for now)
apscheduler==3.10.4

# Validation + settings
pydantic==2.6.1
pydantic-settings==2.1.0

# YAML policy for gateway
pyyaml==6.0.1

# Environment support
python-dotenv==1.0.1

# LLM client
openai==1.12.0

# Logging utils
rich==13.7.0

requests
